agregando sweet alert al eliminar producto del carrito

+++ components/modalCart.js
@@ -3,7 +3,6 @@ import { deleteCartItems } from "../components/cart.js";
 const cartSection = document.querySelector('.cartSection');
 const openCart = document.getElementById('openCart');
 const closeCart = document.getElementById('closeCart');
-const deleteButton = document.querySelector('.delete-button');
 
 openCart.addEventListener('click', () =>{
     cartSection.classList.toggle('cartSectionActive');
@@ -13,28 +12,30 @@ closeCart.addEventListener('click', () =>{
     cartSection.classList.remove('cartSectionActive');
 });
 
-deleteButton.addEventListener('click', (e) =>{
+cartSection.addEventListener('click', (e) =>{
     e.stopPropagation();
 
-    Swal.fire({
-        title: 'Estas seguro?',
-        text: 'El producto se eliminara del carrito de compras',
-        icon: 'warning',
-        showDenyButton: true,
-        confirmButtonText: 'Si, seguro',
-        denyButtonText: 'Cancelar'
-    }).then((result) => {
-        if(result.isConfirmed) {
-            deleteItems();
-            Swal.fire({
-                title: 'Producto eliminado',
-                text: 'El producto ha sido eliminado del carrito de compras',
-                icon: 'success',
-                showConfirmButton: false,
-                timer: 1500
-            });
-        };
-    });
+    if (e.target.classList.contains('delete-button')){
+        Swal.fire({
+            title: 'Estas seguro?',
+            text: 'El producto se eliminara del carrito de compras',
+            icon: 'warning',
+            showDenyButton: true,
+            confirmButtonText: 'Si, seguro',
+            denyButtonText: 'Cancelar'
+        }).then((result) => {
+            if(result.isConfirmed) {
+                deleteItems();
+                Swal.fire({
+                    title: 'Producto eliminado',
+                    text: 'El producto ha sido eliminado del carrito de compras',
+                    icon: 'success',
+                    showConfirmButton: false,
+                    timer: 1500
+                });
+            };
+        });
+    }
     
     // OTRO OPERADOR LOGICO AND
     function deleteItems(){
